Initialize destination registers when needed
Some destination registers, like in the M2_vcmpy_s0_sat_r are used with
the deposit tinycode instruction which needs to read the value of the
register, before writing into it. This caused an uninitialized read
which resulted in a segmentation fault. The issue was addressed by
keeping track of all the declared register, and initializing all the
ones which are declared but not initialized by the macros.