Changed strictness of C compiler
Now using `-Wall, -Wextra` which produced warnings about sign-compare
and unused parameters. These warnings then fixed. The other notable
change is the adding of `%option never-interactive` which removes
the dependency on the undefined `fileno` posix function which is used
when checking whether or not the parser is being run in a tty. This change
is not a problem as we don't use the interactive mode.