Changed back from `-std=c11` to `-std=gnu99`
Also removed duplicate flags such as `-Wall`. The change to `gnu99` is
done to comply with QEMU coding standards.

This resulted in the removal of `%option never-interactive` as `fileno`
is declared in `gnu99` which is not the case in `c11`.

Changing to `gnu99` also removes the `_Generic` macro from `c11`,
thankfully QEMU has their own `QEMU_GENERIC` macro to take its
place.