Fixed indentation and removed some unused params
Indentation of `(void) ...` are now correct in the functions, and the
`reg_arg_print` function was made `static` as it is local to
`parser-helpers.c`. Lastly the `num` parameter is removed from the
`gen_load/store` functions as array arguments aren't handled, and
`num == 1` is asserted at the call sites in `idef-parser.y`.