Redeclared fHIDE macro and added lexing/parsing
The new expressions that were previously hidden in fHIDE are now lexed
and parsed. This primarily includes variable declarations of integer
type such as `int j;`, but also some `if` statements and calls to
a `C`-function `fatal(...)`. This function is currently NOP'd.

Nothing is done with the additional type information yet. Declarations
and assigments such as `int j = 0` are currently treated as `j = 0`
where the type of `j` is the default signed 32-bit int. This results in
two failing tests that use `fSETHALF(n, tmp, ...)` where `tmp = 0` is
now included which sets `tmp` as 32-bit and not 64-bit as `fSETHALF`
would initialize it.  This is easily fixed in the future by including
the new type information.

Signed-off-by: Anton Johansson <anjo@rev.ng>