Now uses the parsed type info to set sign/width
Allocates a new varid of specified sign/width when a variable is
declared.

Now also enforcing signedness in `gen_bin_op` for binary expressions
with variables. This probably needs to be done in more places.

Other minor changes includes setting default signedness and width of
VAR_ID tokens to `signedness = UNKNOWN_SIGNEDNESS` and `bit_width = 0`
to ensure we set all types via declarations. The sign of the immediate
`i` was also set to `SIGNED`.

Signed-off-by: Anton Johansson <anjo@rev.ng>