Added checks for variable declarations
This was done by asserting that a variable exists when lookups are done
using `find_variable`. All variables are now asserted to be declared
except for the effective address `EA`. The variable `EA` is declared
when it appears in an assignment, as previously. A better way to this
might be to declare it as a 32-bit int whenever it appears in the
argument list of a function.

Signed-off-by: Anton Johansson <anjo@rev.ng>