Changed string -> int conversion function in lexer
Now uses `strtoull` for all literals since the sign of the int is
handled explicitly in the parser using the `-` token. By reading in all
ints as unsigned 64-bit we avoid accidental casting bugs.

Also now uses a base of `0` in `strtoull` instead of explicitly setting base `10` and
`16`. Base `0` encompass these by using base 16 when the string starts
with `0x`, base 8 if it starts with `0`, and base 10 otherwise.

Signed-off-by: Anton Johansson <anjo@rev.ng>