accel/tcg: Make uses of TARGET_LONG_BITS target independent
For uses that really need access to the size of a target long, a
long_bits field is added to TCGCPUParams which can be accessed via
CPUClass.

Uses of TCG_OVERSIZED_GUEST which depends on TARGET_LONG_BITS are
also replaced in this commit.

Finally, in decode_sleb128() results are now sign extended to
TARGET_PAGE_BITS instead of TARGET_LONG_BITS, as only page offsets are
encoded. An assert is added to confirm this is always the case.

Signed-off-by: Anton Johansson <anjo@rev.ng>