helper-to-tcg: Introduce PrepareForTcgPass
Adds a new pass over the LLVM module which runs post-optimization with
the end-goal of:

  * Culling functions which aren't worth translating;
  * Canonicalizing the IR to something closer to TCG, and;
  * Extracting information which may be useful in the backend pass.

The bulk of IR transformations occur in the canonicalization phase, with
a handful occuring later in the packend. This commits sets up a new LLVM
pass over the IR module and runs it from the pipeline.

Signed-off-by: Anton Johansson <anjo@rev.ng>