helper-to-tcg: Introduce TcgGenPass
Adds a final backend pass over the LLVM module, taking
previously optimized and canonicalized LLVM IR and for
each function:

  1. Maps arguments, and propagates constant expressions;
  2. Allocates TCG values to non-constants;
  3. Emits the final TCG.

Points above will be implemented in following commits, this commits adds
boilerplate and some scaffolding for translating and emitting code.

TODO: This commit could do with some more splitting, TcgEmit.hpp and
ValueMapping.hpp and the functions therein are referenced but not added
in this commit, they come later.

Signed-off-by: Anton Johansson <anjo@rev.ng>