accel/tcg: Add getpc helper
Introduces a helper function to return the current pc (returnaddress of
the helper), this allows helper-to-tcg to correctly translate nested
helper functions that checks for faulting memory operations.

This is useful as an optimization where a helper functions has a
commonly taken fast path that doesn't fault e.g.

  void HELPER(outer)(...)
  {
      ... // non faulting operations
      if (some_uncommon_condition) {
          helper_inner(..., GETPC())
      }
  }

the outer helper along with the condition can then be emitted as TCG,
and helper_inner() will be emitted as a gen_helper_inner(, ra), where ra
is the result of gen_helper_getpc().

NOTE: This is not ideal since we're introducing extra helper calls,
      and the solution doesn't deal with the "inner" function not being
      a helper.  A better solution and what we'll probably do in the
      next version of the patchset is to instead emit a helper
      definition for the "inner" function that uses GETPC().

Signed-off-by: Anton Johansson <anjo@rev.ng>